+++ src/helper.ts
@@ -1,148 +1,17 @@
 'use strict';
 
 import * as vscode from 'vscode';
+import { CellType, CellAlign, DelimiterType, CellInfo, TableInfo } from './table';
 
 var strWidth = require('string-width')
 var trim = require('trim')
 
-// Cell type
-export enum CellType {
+// Separator type
+enum SeparatorType {
     None,
-    MD_NormalSeparator,
-    MD_LeftSeparator,
-    MD_RightSeparator,
-    MD_CenterSeparator,
-    TT_HeaderPrefix,
-    TT_LeftPrefix,
-    TT_RightPrefix,
-    TT_CenterPrefix
+    Minus,
+    Equall
 };
-
-// Cell align
-export enum CellAlign {
-    Left,
-    Center,
-    Right
-};
-
-// Cell info
-export class CellInfo {
-    private _size: number;
-    private _type: CellType;
-    private _align: CellAlign;
-    private _padding: number;
-
-    constructor(size: number, type: CellType = CellType.None, align: CellAlign = CellAlign.Left, padding: number = 0) {
-        this._size = size;
-        this._type = type;
-        this._align = align;
-        this._padding = padding;
-    }
-
-    get size(): number {
-        return this._size;
-    }
-
-    get type(): CellType {
-        return this._type;
-    }
-
-    get align(): CellAlign {
-        return this._align;
-    }
-
-    get padding(): number {
-        return this._padding;
-    }
-
-    isValid(): boolean {
-        if (this._size < 0) return false;
-        return true;
-    }
-
-    setSize(size: number) {
-        this._size = size;
-    }
-
-    setType(type: CellType) {
-        this._type = type;
-    }
-
-    setAlign(align: CellAlign) {
-        this._align = align;
-    }
-
-    setPadding(padding: number) {
-        this._padding = padding;
-    }
-}
-
-export class TableInfo {
-    private _range: vscode.Range;
-    private _cellGrid: Array<Array<CellInfo>>;
-    private _size: { row: number, col: number };
-
-    constructor(range: vscode.Range, cells: Array<Array<CellInfo>>) {
-        this._range = range;
-        this._cellGrid = cells;
-        this._size = this.getSize();
-    }
-
-    get range(): vscode.Range {
-        return this._range;
-    }
-
-    get cellGrid(): Array<Array<CellInfo>> {
-        return this._cellGrid;
-    }
-
-    get size(): { row: number, col: number } {
-        return this._size;
-    }
-
-    isValid(): boolean {
-        if (!this._range || this._range.isEmpty) return false;
-        if (!this._cellGrid || this._cellGrid.length == 0) return false;
-        var size = this._cellGrid[0].length;
-        this._cellGrid.forEach(row => {
-            if (!row || row.length == 0 || row.length != size) return false;
-            row.forEach(cell => {
-                if (!cell.isValid()) return false;
-            });
-        });
-        return true;
-    };
-
-    private getSize(): { row: number, col: number } {
-        if (!this.isValid()) return { row: 0, col: 0 };
-
-        return {
-            row: this._cellGrid.length,
-            col: this._cellGrid[0].length
-        };
-    };
-
-    getMaxCellSizeList(): Array<number> {
-        if (!this.isValid()) return [];
-
-        var list = [];
-        for (var c = 0; c < this._size.col; c++) {
-            var max = 0;
-            this._cellGrid.forEach((row, r) => {
-                if (r == 0) max = row[c].size;
-                // Use min in 1st col
-                if (c == 0) {
-                    max = Math.min(max, row[c].size);
-                }
-                else {
-                    max = Math.max(max, row[c].size);
-                }
-            })
-            list.push(max);
-        }
-        return list;
-    };
-}
 
 export class TableHelper {
 
@@ -152,37 +21,57 @@ export class TableHelper {
     dispose() {
     }
 
+    // テーブル記法の行か
     public isTableLine(line: vscode.TextLine): boolean {
         if (line.isEmptyOrWhitespace) return false;
-        var tableLineReg = /\|/;
-        return tableLineReg.test(line.text);
+        // 行が"|"を含む、または行が+-=のみで構成されている
+        // @TODO: SimpleTable用の判定
+        return /\||^(?=.*?\+)[\-=+]+$/.test(line.text);
     }
 
+    // 行の解析
     public getCellInfoList(line: vscode.TextLine): Array<CellInfo> {
         if (line.isEmptyOrWhitespace) return [];
 
         var list: Array<CellInfo> = [];
-        var cells = line.text.split("|", -1);
+        var cells = [];
+        var delimiter = DelimiterType.Pipe;
 
-        // 1st element is white space. 
-        list.push(new CellInfo(line.firstNonWhitespaceCharacterIndex));
-        
+        // |がないときのみ、+で分ける
+        if (line.text.indexOf('|') != -1) {
+            cells = line.text.split("|", -1);
+            delimiter = DelimiterType.Pipe;
+        }
+        else {
+            cells = line.text.split("+", -1);
+            delimiter = DelimiterType.Plus;
+        }
+
+        // 空白列を追加
+        list.push(new CellInfo(line.firstNonWhitespaceCharacterIndex, delimiter));
+
         for (var i = 0; i < cells.length; i++) {
             var trimed = trim(cells[i]);
-            // 1st white space
+
+            // 最初の空白は追加済みなので無視する
             if (i == 0 && trimed == "") continue;
-            // last white space
+            // 最後の空白は追加せず無視する
             if (i == cells.length - 1 && trimed == "") continue;
 
             var size = strWidth(trimed);
-            var type = CellType.None;
+            var type = (size == 0) ? CellType.CM_Blank : CellType.CM_Content;
             var align = CellAlign.Left;
 
-            // Markdown ----------------
+            // Common  ----------------
             if (/^-+$/.test(trimed)) {
-                type = CellType.MD_NormalSeparator;
+                type = CellType.CM_MinusSeparator;
                 align = CellAlign.Left;
             }
+            else if (/^=+$/.test(trimed)) {
+                type = CellType.CM_EquallSeparator;
+                align = CellAlign.Left;
+            }
+            // Markdown ----------------
             else if (/^:-+$/.test(trimed)) {
                 type = CellType.MD_LeftSeparator;
                 align = CellAlign.Left;
@@ -217,84 +106,94 @@ export class TableHelper {
                 align = CellAlign.Center;
             }
 
-            list.push(new CellInfo(size, type, align));
+            list.push(new CellInfo(size, delimiter, type, align));
         }
         return list;
     }
-    
-    public getTableInfo(doc: vscode.TextDocument, line: number): TableInfo {
-        // Scan back
-        var startLine = line;
-        for (var i = line; i >= 0; i--) {
-            if (!this.isTableLine(doc.lineAt(i))) break;
-            startLine = i;
-        }
 
-        // Scan forward
-        var endLine = line;
-        for (var i = line + 1; i < doc.lineCount; i++) {
-            if (!this.isTableLine(doc.lineAt(i))) break;
-            endLine = i;
-        }
-
-        // get cell info
-        var grid: Array<Array<CellInfo>> = [];
-        for (var i = startLine; i <= endLine; i++) {
-            grid.push(this.getCellInfoList(doc.lineAt(i)));
-        }
-
-        return new TableInfo(
-            new vscode.Range(
-                new vscode.Position(startLine, 0),
-                new vscode.Position(endLine, doc.lineAt(endLine).range.end.character)
-            ),
-            this.getNormalizedCellGrid(grid)
-        );
-    }
-
+    // 表データの正規化
     private getNormalizedCellGrid(grid: Array<Array<CellInfo>>): Array<Array<CellInfo>> {
-        // Normalize grid size
+        // 全行のサイズを揃える
         var max = 0;
         grid.forEach(row => {
             max = Math.max(max, row.length);
         });
         grid.forEach(row => {
+            // デリミタも揃える
+            var delimiter = (row.length > 0) ? row[0].delimiter : DelimiterType.Pipe;
             for (var c = row.length; c < max; c++) {
-                row.push(new CellInfo(0));
+                row.push(new CellInfo(0, delimiter));
             }
         });
 
-        // Markdown separator
+        // セパレータタイプを確定する
         grid.forEach(row => {
-            var isSeparatorRow = false;
-            row.forEach(cell => {
-                if (cell.type == CellType.MD_NormalSeparator || cell.type == CellType.MD_LeftSeparator || cell.type == CellType.MD_RightSeparator || cell.type == CellType.MD_CenterSeparator) {
-                    isSeparatorRow = true;
+            // セパレータ行かの判定
+            var rowType = SeparatorType.None;
+            for (var i = 0; i < row.length; i++) {
+                var cell = row[i];
+                // デリミタがPlusなら初期値をMinusにする（一応この時点でセパレータ行で確定ではあるが特に何もしない）
+                if (i == 0 && cell.delimiter == DelimiterType.Plus) {
+                    rowType = SeparatorType.Minus;
+                }
+
+                // 文字列なら非セパレータ行で確定
+                if (cell.type == CellType.CM_Content) {
+                    rowType = SeparatorType.None;
+                    break;
+                }
+
+                // セルのタイプで判定
+                if (cell.type == CellType.CM_MinusSeparator) {
+                    rowType = SeparatorType.Minus;
+                }
+                else if (cell.type == CellType.CM_EquallSeparator) {
+                    rowType = SeparatorType.Equall;
+                }
+                else if (cell.type == CellType.MD_LeftSeparator || cell.type == CellType.MD_RightSeparator || cell.type == CellType.MD_CenterSeparator) {
+                    rowType = SeparatorType.Minus;
+                }
+            }
+
+            // セパレータタイプを補正
+            row.forEach((cell, i) => {
+                // セパレータ行でない場合、セルタイプを文字列に（-や:-を文字として扱う）
+                if (rowType == SeparatorType.None) {
+                    if (cell.type == CellType.CM_MinusSeparator || cell.type == CellType.CM_EquallSeparator ||
+                        cell.type == CellType.MD_LeftSeparator || cell.type == CellType.MD_RightSeparator || cell.type == CellType.MD_CenterSeparator) {
+                        cell.setType(CellType.CM_Content);
+                    }
+                }
+                // セパレータ行の場合、セルタイプが空白のものをセパレータに
+                else {
+                    if (i != 0 && cell.type == CellType.CM_Blank) {
+                        switch (rowType) {
+                            case SeparatorType.Minus:
+                                cell.setType(CellType.CM_MinusSeparator);
+                                break;
+                            case SeparatorType.Equall:
+                                cell.setType(CellType.CM_EquallSeparator);
+                                break;
+                        }
+                    }
                 }
             });
-            // Set separator if text is white space
-            if (isSeparatorRow) {
-                row.forEach((cell, i) => {
-                    if (i != 0 && cell.size == 0) {
-                        cell.setType(CellType.MD_NormalSeparator);
-                    }
-                });
-            }
         });
 
-        // Markdown size
+        // Markdownのセパレータのサイズを設定
         grid.forEach(row => {
             row.forEach(cell => {
-                // Minimum size is 3
-                if (cell.type == CellType.MD_NormalSeparator || cell.type == CellType.MD_LeftSeparator || cell.type == CellType.MD_RightSeparator || cell.type == CellType.MD_CenterSeparator) {
+                if (cell.type == CellType.CM_MinusSeparator ||
+                    cell.type == CellType.MD_LeftSeparator || cell.type == CellType.MD_RightSeparator || cell.type == CellType.MD_CenterSeparator) {
+                    // 最小である3文字にする
                     cell.setSize(3);
                 }
             });
         });
 
-        // Textile size
+        // Textileのサイズを設定
         grid.forEach(row => {
-            // Set padding for prefix
+            // プレフィックス分のパディングを他の行の同列に設定する
             row.forEach((cell, i) => {
                 if (cell.type == CellType.TT_HeaderPrefix || cell.type == CellType.TT_LeftPrefix || cell.type == CellType.TT_RightPrefix || cell.type == CellType.TT_CenterPrefix) {
                     grid.forEach(elem => {
@@ -306,41 +205,41 @@ export class TableHelper {
             });
         });
 
-        // Markdown align
+        // Markdownの位置揃え
         for (var r = grid.length - 1; r >= 0; r--) {
             var row = grid[r];
+            // 各列の位置揃えを変更する
             row.forEach((cell, c) => {
-                // Set col align
                 switch (cell.type) {
                     case CellType.MD_LeftSeparator:
                         grid.forEach(elem => {
                             if (c < elem.length) {
                                 elem[c].setAlign(CellAlign.Left);
                             }
-                        });    
+                        });
                         break;
                     case CellType.MD_RightSeparator:
                         grid.forEach(elem => {
                             if (c < elem.length) {
                                 elem[c].setAlign(CellAlign.Right);
                             }
-                        });    
+                        });
                         break;
                     case CellType.MD_CenterSeparator:
                         grid.forEach(elem => {
                             if (c < elem.length) {
                                 elem[c].setAlign(CellAlign.Center);
                             }
-                        });    
+                        });
                         break;
                 }
             });
         }
 
-        // Textile align
+        // Textileの位置揃え
         grid.forEach(row => {
+            // 各セルの位置揃えを変更する
             row.forEach(cell => {
-                // Set cell align
                 switch (cell.type) {
                     case CellType.TT_LeftPrefix:
                         cell.setAlign(CellAlign.Left)
@@ -356,5 +255,36 @@ export class TableHelper {
         });
 
         return grid;
+    }
+
+    // 表データの取得
+    public getTableInfo(doc: vscode.TextDocument, line: number): TableInfo {
+        // 後方に操作し開始行を取得
+        var startLine = line;
+        for (var i = line; i >= 0; i--) {
+            if (!this.isTableLine(doc.lineAt(i))) break;
+            startLine = i;
+        }
+
+        // 前方に操作し終了行を取得
+        var endLine = line;
+        for (var i = line + 1; i < doc.lineCount; i++) {
+            if (!this.isTableLine(doc.lineAt(i))) break;
+            endLine = i;
+        }
+
+        // 各行の解析
+        var grid: Array<Array<CellInfo>> = [];
+        for (var i = startLine; i <= endLine; i++) {
+            grid.push(this.getCellInfoList(doc.lineAt(i)));
+        }
+
+        return new TableInfo(
+            new vscode.Range(
+                new vscode.Position(startLine, 0),
+                new vscode.Position(endLine, doc.lineAt(endLine).range.end.character)
+            ),
+            this.getNormalizedCellGrid(grid)
+        );
     }
 }