+++ CHANGELOG.md
@@ -1,7 +1,25 @@
 # Change Log
-All notable changes to the "vscodetableformatter" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-- Initial release
+
+* reStructuredText
+    * Simple table
+
+## [1.0.1] - 2017-
+
+* Support reStructuredText table syntax.
+    * Grid table
+
+## [1.0.0] - 2017-01-22
+
+### Added
+
+* Support Markdown table syntax.
+    * Align
+    * Table header separator
+* Support Textile table syntax.
+    * Align
+    * Table header
+* Add command.
+    * Table: Foramt All
+    * Table: Foramt Current